Fix loading workflows with explicit reader settings
In the owfile settings, explicitly selected selected readers are stored with
full module and class name. Moving but not adding imports showed
the missing reader error.
